services/notifications/requirements.txt
# FastAPI and Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.20

# Database
sqlmodel==0.0.22
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Dapr for Event-Driven Architecture
dapr==1.14.0
dapr-ext-fastapi==1.14.0

# Configuration and Environment
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# HTTP Client
httpx==0.28.1

# Date/Time Utilities
python-dateutil==2.9.0.post0

# Email
aiosmtplib==3.0.2

# Logging and Monitoring
# (Using Python standard library for logging)

# Testing (development)
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
